Refactor the detector component into better structure
