+++ sequence/components/detector/qs_timebin_detector.py
@@ -1,0 +1,68 @@
+from typing import TYPE_CHECKING, Any
+
+from .base import QSDetector, Detector
+from ..interferometer import Interferometer
+from ..switch import Switch
+from ...utils.encoding import time_bin
+
+if TYPE_CHECKING:
+    from sequence.kernel.timeline import Timeline
+
+class QSDetectorTimeBin(QSDetector):
+    """QSDetector to measure time bin encoded qubits.
+
+    There are three detectors.
+    The switch is connected to detectors[0] and the interferometer.
+    The interferometer is connected to detectors[1] and detectors[2].
+
+    Attributes:
+        name (str): label for QSDetector instance.
+        timeline (Timeline): timeline for simulation.
+        detectors (list[Detector]): list of attached detectors (length 3).
+        trigger_times (list[list[int]]): tracks simulation time of detection events for each detector.
+        switch (Switch): internal optical switch component.
+        interferometer (Interferometer): internal interferometer component.
+    """
+
+    def __init__(self, name: str, timeline: "Timeline"):
+        QSDetector.__init__(self, name, timeline)
+        self.switch = Switch(name + ".switch", timeline)
+        self.detectors = [Detector(name + ".detector" + str(i), timeline) for i in range(3)]
+        self.switch.add_receiver(self.detectors[0])
+        self.interferometer = Interferometer(name + ".interferometer", timeline, time_bin["bin_separation"])
+        self.interferometer.add_receiver(self.detectors[1])
+        self.interferometer.add_receiver(self.detectors[2])
+        self.switch.add_receiver(self.interferometer)
+
+        self.components = [self.switch, self.interferometer] + self.detectors
+        self.trigger_times = [[], [], []]
+
+    def init(self):
+        """Implementation of Entity interface (see base class)."""
+
+        assert len(self.detectors) == 3
+        super().init()
+
+    def get(self, photon, **kwargs):
+        """Method to receive a photon for measurement.
+
+        Forwards the photon to the internal fiber switch.
+
+        Args:
+            photon (Photon): photon to measure.
+
+        Side Effects:
+            Will call `get` method of attached switch.
+        """
+
+        self.switch.get(photon)
+
+    def get_photon_times(self):
+        times, self.trigger_times = self.trigger_times, [[], [], []]
+        return times
+
+    def set_basis_list(self, basis_list: list[int], start_time: int, frequency: float) -> None:
+        self.switch.set_basis_list(basis_list, start_time, frequency)
+
+    def update_interferometer_params(self, arg_name: str, value: Any) -> None:
+        self.interferometer.__setattr__(arg_name, value)
